etl_job/requirements.txt
pymongo==3.12.0
SQLAlchemy
psycopg2-binary==2.9.3
pandas
vaderSentiment